first 50 fibonacci

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -8,9 +8,9 @@
 
 int main(void)
 {
-	unsigned long long int num1 = 1;
-	unsigned long long int num2 = 2;
-	unsigned long long int fib;
+	unsigned long int num1 = 1;
+	unsigned long int num2 = 2;
+	unsigned long int fib;
 	int i;
 
 	printf("%lu, ", num1);
